+++ org/security/list-users-by-MFA-methods-count.ps1
@@ -1,4 +1,4 @@
-﻿<#
+<#
     .SYNOPSIS
     Report users by the count of their registered MFA methods
 
@@ -35,7 +35,7 @@ RunbookCustomization: {
 #>
 
 #Requires -Modules @{ModuleName = "RealmJoin.RunbookHelper"; ModuleVersion = "0.8.6" }
-#Requires -Modules @{ModuleName = "Microsoft.Graph.Authentication"; ModuleVersion = "2.35.1" }
+#Requires -Modules @{ModuleName = "Microsoft.Graph.Authentication"; ModuleVersion = "2.37.0" }
 
 param (
     [Parameter(Mandatory = $true)]
@@ -56,7 +56,7 @@ if ($CallerName) {
     Write-RjRbLog -Message "Caller: '$CallerName'" -Verbose
 }
 
-$Version = "1.0.0"
+$Version = "1.0.1"
 Write-RjRbLog -Message "Version: $Version" -Verbose
 
 # Add Parameter in Verbose output